Don't pass int to %f printf format
Assuming %f was intended not %d

+++ src/Lib/DEM/dem.cxx
@@ -446,8 +446,8 @@ TGDem::write_area( const string& root, SGBucket& b ) {
     }
 
     gzprintf( fp, "%d %d\n", (int)min_x, (int)min_y );
-    gzprintf( fp, "%d %f %d %f\n", span_x + 1, (int)col_step,
-              span_y + 1, (int)row_step );
+    gzprintf( fp, "%d %f %d %f\n", span_x + 1, col_step,
+              span_y + 1, row_step );
     for ( int i = start_x; i <= start_x + span_x; ++i ) {
         for ( int j = start_y; j <= start_y + span_y; ++j ) {
             gzprintf( fp, "%d ", (int)dem_data[i][j] );
